Bring proto on-par with RFC

+++ encodings/otlp/generator.go
@@ -34,8 +34,8 @@ func (g *Generator) GenerateBatch(spansPerBatch int, attrsPerSpan int, timedEven
 
 	resource := Resource{
 		Labels: []*AttributeKeyValue{
-			{Key: "StartTimeUnixnano", IntValue: 12345678},
-			{Key: "Pid", IntValue: 1234},
+			{Key: "StartTimeUnixnano", Int64Value: 12345678},
+			{Key: "Pid", Int64Value: 1234},
 			{Key: "HostName", StringValue: "fakehost"},
 			{Key: "ServiceName", StringValue: "generator"},
 		},
@@ -64,9 +64,9 @@ func (g *Generator) GenerateBatch(spansPerBatch int, attrsPerSpan int, timedEven
 
 			if attrsPerSpan >= 2 {
 				span.Attributes = append(span.Attributes,
-					&AttributeKeyValue{Key: "load_generator.span_seq_num", Type: AttributeKeyValue_INT64, IntValue: int64(spanID)})
+					&AttributeKeyValue{Key: "load_generator.span_seq_num", Type: AttributeKeyValue_INT64, Int64Value: int64(spanID)})
 				span.Attributes = append(span.Attributes,
-					&AttributeKeyValue{Key: "load_generator.trace_seq_num", Type: AttributeKeyValue_INT64, IntValue: int64(traceID)})
+					&AttributeKeyValue{Key: "load_generator.trace_seq_num", Type: AttributeKeyValue_INT64, Int64Value: int64(traceID)})
 			}
 
 			for j := len(span.Attributes); j < attrsPerSpan; j++ {
@@ -81,7 +81,7 @@ func (g *Generator) GenerateBatch(spansPerBatch int, attrsPerSpan int, timedEven
 				span.TimedEvents = append(span.TimedEvents, &Span_TimedEvent{
 					TimeUnixnano: core.TimeToTimestamp(startTime),
 					Attributes: []*AttributeKeyValue{
-						{Key: "te", Type: AttributeKeyValue_INT64, IntValue: int64(spanID)},
+						{Key: "te", Type: AttributeKeyValue_INT64, Int64Value: int64(spanID)},
 					},
 				})
 			}
